ebook_builder.sh: By default don't include cover image in EPUB.
It's the single biggest contributor to the size of the AYiV EPUB, at least
after illustrations at unreasonably-large resolutions are fixed, and the
Kindle guidelines say not to include the cover image embedded in the source
file. Removing it gets us out of the ebook-size warning message on the Kindle
Store page.