Fix bad indexing when first line contains exclamation mark.
Thanks to Rachel Newhouse for pointing the problem out. (The exclamation
mark turned into an index sub-entry---a feature which I make use of
occasionally in my books.) It took awhile to figure out that the problem
was in the TeX output by md2tex.sh, and then yet longer to figure out
that simply applying the index-quote character in the command caused it
to also display in the text (a worse problem) and a *proper* fix, which
required two new commands in poemscompat.tex and a change to md2tex.sh
to use them, but everything looks good now.